refactor: CustomDateFomatter의 내용을  DateFormatter의 extension으로 구현
- Extension 그룹에 DateFormatter 파일 추가
- CustomDateFormatter 삭제
- MainVCTableViewDataSource에 CustomDateFomatter로 구현한 내용 수정
